third_party/WebKit/LayoutTests/dom/text/next-element-sibling.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description('This tests the NonDocumentTypeChildNode interface nextElementSibling property on a Text node.');

var element = document.createElement('div');
var a = element.appendChild(new Text('a'));
shouldBeNull('a.nextElementSibling');
var b = element.appendChild(document.createElement('b'));
shouldBe('a.nextElementSibling', 'b');
var c = element.insertBefore(new Comment('c'), b);
shouldBe('a.nextElementSibling', 'b');

</script>
