third_party/blink/web_tests/dom/text/next-element-sibling-expected.txt

This tests the NonDocumentTypeChildNode interface nextElementSibling property on a Text node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS a.nextElementSibling is null
PASS a.nextElementSibling is b
PASS a.nextElementSibling is b
PASS successfullyParsed is true

TEST COMPLETE
